refractor: prices filtered by stream

+++ src/main/java/com/app/infrastructure/outputport/PriceRepository.java
@@ -9,7 +9,7 @@ import java.time.LocalDateTime;
 import java.util.List;
 
 @Repository
-public interface PriceRepository extends JpaRepository<Price, Integer> {
+public interface PriceRepository extends JpaRepository<Price, Long> {
 
     List<Price> findPrices(LocalDateTime applicationRange, int productId, int brandId);
